+++ packages/radix-ui-themes/changelog.md
@@ -38,6 +38,13 @@
   - Remove comments from the CSS build.
   - Make sure that forced light/dark appearance on the `Theme` component also sets the corresponding browser colors, like the correct input autofill background color.
   - Use `outline` rather than `box-shadow` for most focus styles, which avoids a slight anti-aliasing issue in Chrome on focused elements.
+  - Add CSS variables to control the cursor style on interactive elements:
+    - `--cursor-link: pointer;`
+    - `--cursor-button: default;`
+    - `--cursor-menu: default;`
+    - `--cursor-disabled: not-allowed;`
+  - Replace `.rt-reset-button` and `.rt-reset-a` classes with a single `.rt-reset` class
+    - **[Upgrade guide]** If you were using these classes for your custom components, update the class name used.
 - `Button`, `IconButton`
   - Refine and normalise the look and feel of the disabled states.
   - Improve `variant="classic"` look and feel across different accent colors in light and dark mode.